Use some refactored functions in server mode's tests
